+++ src/Miningcore/DataStore/Postgres/Scripts/db_change_20211004.sql
@@ -3,6 +3,6 @@ DROP TABLE IF EXISTS payment_schedule;
 CREATE TABLE poolstate
 (
 	poolid TEXT NOT NULL PRIMARY KEY,
-	hashvalue DOUBLE PRECISION NOT NULL DEFAULT 0,
+	hashvalue DECIMAL NOT NULL DEFAULT 0,
 	lastpayout TIMESTAMP NULL
 );
